Do not use functional cast on integers.
GitOrigin-RevId: 1ebcf0e9a9b004039d32b6363140387bca7d7722

+++ td/mtproto/Transport.cpp
@@ -337,7 +337,7 @@ size_t Transport::write_no_crypto(const Storer &storer, PacketInfo *info, Mutabl
     return size;
   }
   // NoCryptoHeader
-  as<uint64>(dest.begin()) = uint64(0);
+  as<uint64>(dest.begin()) = 0;
   auto real_size = storer.store(dest.ubegin() + sizeof(uint64));
   CHECK(real_size == storer.size());
   return size;
